match ChromeDriver version to installed Chrome package
I originally tried to hack the Dockerfile to use some kind of variable
in the wget command, but it doesn't seem to be possible, so instead I
refactored the Chrome driver installation into its own shell script.

This also requires installing Chrome beta instead of unstable, as there
is currently only a ChromeDriver version for beta.